Add ok and some pipe methods
`some` or `some_fn` expects an Option<T> result. The result is stored
in the pipeline. If `T` is none, the flow will be stopped.

`ok` or `ok_fn` expects a Result<T,E> result. The result is stored in
the pipeline. If `T` is `Err`, the flow will be stopped.